+++ Engine/Models/Square_Anim_10x10.h
@@ -16,7 +16,7 @@ static const u32 g_Square_Anim_10x10_numberOfPrimitives = 1;
 static const u32 g_Square_Anim_10x10_numberOfAttributes = 2;
 
 static PrimitiveData g_Square_Anim_10x10_PrimitiveArray[g_Square_Anim_10x10_numberOfPrimitives]={
-	{GL_TRIANGLE_STRIP,(f32*)g_Square_Anim_10x10_VertexData0,NULL,g_Square_Anim_10x10_numberOfVertices,sizeof(g_Square_Anim_10x10_VertexData0),0,0,0},
+	{GL_TRIANGLE_STRIP,(u8*)g_Square_Anim_10x10_VertexData0,NULL,g_Square_Anim_10x10_numberOfVertices,sizeof(g_Square_Anim_10x10_VertexData0),0,0,0},
 };
 
 static AttributeData g_Square_Anim_10x10_AttributesArray[g_Square1x1_numberOfAttributes]=
