+++ Engine/Models/Square_Tiled_16.h
@@ -14,7 +14,7 @@ static const u32 g_Square_Tiled_16_numberOfPrimitives = 1;
 static const u32 g_Square_Tiled_16_numberOfAttributes = 2;
 
 static PrimitiveData g_Square_Tiled_16_PrimitiveArray[g_Square_Tiled_16_numberOfPrimitives]={
-	{GL_TRIANGLE_STRIP,(f32*)g_Square_Tiled_16_VertexData0,NULL,g_Square_Tiled_16_numberOfVertices,sizeof(g_Square_Tiled_16_VertexData0),0,0,0},
+	{GL_TRIANGLE_STRIP,(u8*)g_Square_Tiled_16_VertexData0,NULL,g_Square_Tiled_16_numberOfVertices,sizeof(g_Square_Tiled_16_VertexData0),0,0,0},
 };
 
 static AttributeData g_Square_Tiled_16_AttributesArray[g_Square1x1_numberOfAttributes]=
